requirements.txt
Flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
openai==1.17.0
langchain==0.1.14
langchain-openai==0.0.6
langchain-pinecone==0.0.3
pinecone-client==3.0.0
